forum_backend/requirements.txt
amqp==2.4.2
Babel==2.6.0
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==2.2.4
django-appconf==1.0.3
django-cors-headers==3.0.2
django-environ==0.4.5
django-imagekit==4.0.2
django-ipware==2.1.0
django-model-utils==3.1.2
django-rest-framework==0.1.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
drf-yasg==1.17.1
flower==0.9.3
idna==2.8
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.2
kombu==4.5.0
MarkupSafe==1.1.1
oauthlib==3.0.1
packaging==20.3
pilkit==2.0
Pillow==6.0.0
PyJWT==1.7.1
pyparsing==2.4.7
python-logstash==0.4.6
python3-openid==3.1.0
pytz==2019.1
redis==3.2.1
requests==2.21.0
requests-oauthlib==1.2.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.12.0
some-package==0.1
sqlparse==0.3.0
tornado==5.1.1
uritemplate==3.0.1
urllib3==1.24.2
uWSGI==2.0.18
vine==1.3.0
